Unify top-level weight loading behind TopLevelAssetLoader
One load path for every trunk's top-level weights instead of four:

- loadTestingTopLevelAssets was never testing-only; build() called it.
  It becomes TopLevelAssetLoader.load in its own file, so the name no
  longer lies about what it does.
- generateFromTokenSuiteForTesting carried a verbatim copy of the same
  GPT-2 load switch; it now calls the loader.
- CPUExactLlamaRuntime.init and loadCachedExactCPULlamaWeights both
  re-typed the llama tokenEmbedding/finalNormGamma/lmHead triple; both
  now share TopLevelAssetLoader.loadLlamaCoreWeights. lmHeadFP16 stays
  a separate load for consumers that need it, preserving memory
  behavior for callers that don't.

compileBlobPath and resolveBundleWeightReference become internal so the
loader module can reach them; behavior is unchanged everywhere.
